Improve parent/child movements
These movements now skip nodes which have no siblings. The reason for
this is to avoid landing on a node with no siblings which would
prevent subsequent forward/backward movement.

This is particularly noticeable in Python where a
"function_definition" contains a "body", which contains a "block" (no
siblings), which then contains a series of "expression_statement"
nodes. Previously moving downwards would require a step for each of
these without much in the way of visual feedback, which might be
confusing. Now, the downward movement skips all child nodes until it
reaches the first "expression_statement".

This change has been applied to both upward and downward movements.