Change movement function return type (Symex move or nil)
The movement functions now return either a Symex move type (if a
movement was made) or nil (if no movement was made). This replicates
the return types of the lisp movement primitives.

The move type is a list tagged with the "move" symbol and containing x
and y offsets indicating the number of nodes traversed. The x-axis
represents sibling movement and the y-axis represents parent/child
movement.

For example, a movement from a node to its next sibling twice is
represented by the list `('move 2 0)` and a movement to a node's
parent is `('move 0 -1)`.

The movement always represents the actual relative movement, not
necessarily the desired movement. For example, a desired movement to
the 3rd next sibling would result in a movement of only 2 if there
were only two more siblings.