Fix returned movement when no further movements are possible
`symex-ts--move-with-count` now returns only the movement that could
actually be achieved.

For example when trying to move to the parent of the root node,
`symex-ts--move-with-count` now returns `nil`. Previously it would
return the movement (e.g. `(move 0 -1)`) even though that movement did
not occur.

If multiple movements are requested, such as moving to the parent
twice, then only the number of moves that could be achieved will be
returned. If none can be achieved then this is `nil`, otherwise it
will be the actual movement that occurred.

This was causing infinite loops to occur in Symex.el for traversals
including cycles as it did not know when to abort the cycle.